Modular Webpack config, and Webpack + Gulp integration.
The Gulpfile now stores variants of the Wepback config--currently using anonymous functions, but it works, and it's somewhat elegant (though it may be messy. Meh). It also successfully builds for production and development (should work more on tweaking the configs for those builds, but this is a good place to start), and can run the dev server, which also has its own config.

I've also removed the other files (./src/raw, .src/processed) that were accidentally added in when I merged this version of the project with the old one (just so that I could push the new one to git).

This is the last commit where the current gulpfile will still contain the old ways of working wth Webpack.